tests: Move test to bounded.rs
Previously the force_push test was placed in unbounded.rs by accident
